Refactor out utility functions.
